Updates Spring Boot and Gradle versions
Updates Spring Boot to version 4.0.0 and Gradle to version 8.14.

Removes explicit Spring Boot test dependencies as they are managed by the Spring Boot starter.

Excludes junit vintage engine to avoid conflicts with newer JUnit versions.

Introduces a base integration test class to reduce boilerplate code in integration tests. It sets up the MockMvc environment for each test.
