+++ accounts/build.gradle
@@ -1,6 +1,6 @@
 plugins {
-	id 'org.springframework.boot' version '3.3.5'
-	id 'io.spring.dependency-management' version '1.1.6'
+	id 'org.springframework.boot' version '4.0.0'
+	id 'io.spring.dependency-management' version '1.1.7'
 	id 'java'
 	id 'groovy'
 }
@@ -43,7 +43,9 @@ dependencies {
 	
 	runtimeOnly 'com.h2database:h2'
 	
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
+	testImplementation('org.springframework.boot:spring-boot-starter-test') {
+		exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
+	}
 	testImplementation('org.spockframework:spock-core:2.3-groovy-4.0') {
 		exclude group: 'org.codehaus.groovy'
 	}
@@ -52,11 +54,6 @@ dependencies {
 	}
 	testImplementation 'org.apache.groovy:groovy:4.0.15'
 	testRuntimeOnly 'net.bytebuddy:byte-buddy:1.14.8'
-	
-	// Add explicit Spring Boot test dependencies for Spock
-	testImplementation 'org.springframework:spring-test'
-	testImplementation 'org.springframework.boot:spring-boot-test'
-	testImplementation 'org.springframework.boot:spring-boot-test-autoconfigure'
 }
 
 tasks.named('test') {
